+++ src/main/java/com/ibiz/api/model/BizChanceSearchVO.java
@@ -350,12 +350,12 @@ public class BizChanceSearchVO {
     /**
      * 변경전 매출금액
      */
-    private Integer befSellAmt;
+    private Long befSellAmt;
 
     /**
      * 변경전 매입금액
      */
-    private Integer befBuyAmt;
+    private Long befBuyAmt;
 
     /**
      * 변경전 매출확률금액
@@ -945,16 +945,16 @@ public class BizChanceSearchVO {
     public void setBefFcstYam(String befFcstYam) {
         this.befFcstYam = befFcstYam;
     }
-    public Integer getBefSellAmt() {
+    public Long getBefSellAmt() {
         return befSellAmt;
     }
-    public void setBefSellAmt(Integer befSellAmt) {
+    public void setBefSellAmt(Long befSellAmt) {
         this.befSellAmt = befSellAmt;
     }
-    public Integer getBefBuyAmt() {
+    public Long getBefBuyAmt() {
         return befBuyAmt;
     }
-    public void setBefBuyAmt(Integer befBuyAmt) {
+    public void setBefBuyAmt(Long befBuyAmt) {
         this.befBuyAmt = befBuyAmt;
     }
     public String getBefSellPrbbAmt() {
